Extract business card test into separate file
